+++ app/Http/Controllers/HomeBatikController.php
@@ -69,7 +69,7 @@ class HomeBatikController extends Controller
         $validated = $request->except(['_token', 'batik_picture']);
 
         if ($request->batik_picture) {
-            unlink($batik->batik_picture);
+            if(file_exists(public_path($batik->batik_picture))) unlink($batik->batik_picture);
             $image = $request->file('batik_picture');
             $batik_picture = 'image/' . time() . uniqid() . '.' . $image->getClientOriginalExtension();
             Image::make($image)->resize(770, 450)->save($batik_picture);
@@ -91,7 +91,7 @@ class HomeBatikController extends Controller
     }
 
     public function destroy(Batik $batik) {
-        unlink($batik->batik_picture);
+        if(file_exists(public_path($batik->batik_picture))) unlink($batik->batik_picture);
         $batik->delete();
 
         $notification = [
