+++ app/Http/Controllers/HomeSubCategoryController.php
@@ -70,7 +70,7 @@ class HomeSubCategoryController extends Controller
     public function destroy(SubCategory $subcategory) {
         $batiks = Batik::where('sub_id', $subcategory->id)->get();
         foreach ($batiks as $batik) {
-            unlink($batik->batik_picture);
+            if(file_exists(public_path($batik->batik_picture))) unlink($batik->batik_picture);
         }
         Batik::where('sub_id', $subcategory->id)->delete();
         $subcategory->delete();
